Squashed 'thirdParty/mallocMC/' changes from 6ac89ac502..c82b95f46c
c82b95f46c update to latest alpaka 1.0.0-dev (#242)

git-subtree-dir: thirdParty/mallocMC
git-subtree-split: c82b95f46c4f1bf6375f760bd5e46d22a0d94585

+++ alpaka/script/install_clang.sh
@@ -1,13 +1,8 @@
 #!/bin/bash
 
 #
-# Copyright 2021 Benjamin Worpitz, Bernhard Manfred Gruber
-#
-# This file is part of alpaka.
-#
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+# Copyright 2023 Benjamin Worpitz, Bernhard Manfred Gruber, Jan Stephan
+# SPDX-License-Identifier: MPL-2.0
 #
 
 source ./script/travis_retry.sh
@@ -18,61 +13,59 @@ source ./script/set.sh
 : "${ALPAKA_CI_STDLIB?'ALPAKA_CI_STDLIB must be specified'}"
 : "${CXX?'CXX must be specified'}"
 
-# add clang-11 repository for ubuntu 18.04
-if [[ "$(cat /etc/os-release)" == *"18.04"* && "${ALPAKA_CI_CLANG_VER}" -eq 11 ]]
+#TODO(SimeonEhrig): remove this statement, if ppa's are fixed in alpaka-group-container
+if [[ -f "/etc/apt/sources.list.d/llvm.list" ]];
 then
-    travis_retry sudo DEBIAN_FRONTEND=noninteractive apt-get -y --quiet --allow-unauthenticated --no-install-recommends install tzdata
-    travis_retry apt-get -y --quiet install wget gnupg2
-    wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
-    echo "deb http://apt.llvm.org/bionic/ llvm-toolchain-bionic-11 main" | sudo tee /etc/apt/sources.list.d/clang11.list
-    echo "deb-src http://apt.llvm.org/bionic/ llvm-toolchain-bionic-11 main" | sudo tee -a  /etc/apt/sources.list.d/clang11.list
-    travis_retry apt-get -y --quiet update
+    sudo rm /etc/apt/sources.list.d/llvm.list
 fi
 
-# add clang-13 repository for ubuntu 20.04
-if [[ "$(cat /etc/os-release)" == *"20.04"* && "${ALPAKA_CI_CLANG_VER}" -eq 13 ]]
+if ! agc-manager -e clang@${ALPAKA_CI_CLANG_VER}
 then
+    # Install from LLVM repository (if available); otherwise install LLVM from official Ubuntu repositories
+    ALPAKA_CI_UBUNTU_NAME=`lsb_release -c | awk '{print $2}'`
     wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
-    sudo add-apt-repository 'deb http://apt.llvm.org/focal/ llvm-toolchain-focal-13 main'
-fi
 
-travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install clang-${ALPAKA_CI_CLANG_VER}
-
-if [ "${ALPAKA_CI_STDLIB}" == "libc++" ]
-then
-    travis_retry sudo apt-get -y --quiet update
-    if [ "${ALPAKA_CI_CLANG_VER}" -gt 6 ]
+    # focal = 20.04; jammy = 22.04
+    if { [ "${ALPAKA_CI_UBUNTU_NAME}" == "focal" ] && [ "${ALPAKA_CI_CLANG_VER}" -ge 13 ]; } || \
+    { [ "${ALPAKA_CI_UBUNTU_NAME}" == "jammy" ] && [ "${ALPAKA_CI_CLANG_VER}" -ge 15 ]; }
     then
+        sudo add-apt-repository "deb http://apt.llvm.org/${ALPAKA_CI_UBUNTU_NAME}/ llvm-toolchain-${ALPAKA_CI_UBUNTU_NAME}-$ALPAKA_CI_CLANG_VER main"
+    fi
+
+    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install clang-${ALPAKA_CI_CLANG_VER}
+
+    if [ -n "${ALPAKA_CI_SANITIZERS}" ]
+    then
+        # llvm-symbolizer is required for meaningful output. This is part of the llvm base package which we don't install by default.
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install llvm-${ALPAKA_CI_CLANG_VER}
+
+        # The sanitizer libraries are part of libclang-rt-${ALPAKA_CI_CLANG_VER}-dev
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libclang-rt-${ALPAKA_CI_CLANG_VER}-dev
+    fi
+
+    if [ "${ALPAKA_CI_STDLIB}" == "libc++" ]
+    then
+        travis_retry sudo apt-get -y --quiet update
         travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++-${ALPAKA_CI_CLANG_VER}-dev
         travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++abi-${ALPAKA_CI_CLANG_VER}-dev
-    else
-        # Ubuntu started numbering libc++ with version 7. If we got to this point, we need to install the 
-        # default libc++ and hope for the best
-        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++-dev
-        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++abi-dev
+        if [ "${ALPAKA_CI_CLANG_VER}" -ge 12 ]
+        then
+            # Starting from LLVM 12 libunwind is required when using libc++. For some reason this isn't installed by default
+            travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libunwind-${ALPAKA_CI_CLANG_VER}-dev
+        fi
     fi
-fi
 
-if [ "${alpaka_ACC_CPU_B_OMP2_T_SEQ_ENABLE}" = "ON" ] || [ "${alpaka_ACC_CPU_B_SEQ_T_OMP2_ENABLE}" = "ON" ] || [ "${alpaka_ACC_ANY_BT_OMP5_ENABLE}" = "ON" ]
-then
-    if [[ "${ALPAKA_CI_CLANG_VER}" =~ ^[0-9]+$ ]] && [ "${ALPAKA_CI_CLANG_VER}" -ge 8 ]
+    if [ "${alpaka_ACC_CPU_B_OMP2_T_SEQ_ENABLE}" = "ON" ] || [ "${alpaka_ACC_CPU_B_SEQ_T_OMP2_ENABLE}" = "ON" ]
     then
         LIBOMP_PACKAGE=libomp-${ALPAKA_CI_CLANG_VER}-dev
-    else
-        LIBOMP_PACKAGE=libomp-dev
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install "${LIBOMP_PACKAGE}"
     fi
-    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install "${LIBOMP_PACKAGE}"
-    if [ "${alpaka_ACC_ANY_BT_OMP5_ENABLE}" = "ON" ]
-    then
-        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install \
-            clang-tools-${ALPAKA_CI_CLANG_VER} llvm-${ALPAKA_CI_CLANG_VER}
-    fi
-fi
 
-sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
-sudo update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
-sudo update-alternatives --install /usr/bin/cc cc /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
-sudo update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/cc cc /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
+fi
 
 which "${CXX}"
 ${CXX} -v
